Merged in OBSREVE-1763-tiered-cache (pull request #892)
OBSRVE-1763 [WIP] Two tier priority cache for trace data

* tiered priority cache implementation

small improves

* integrate priority caching into the main processor

* change behaviour of root_spans evaluator, so that it return LowPriority when a lonely root span occurs

Instead of returning "NotSampled" it will return the softer LowPriority. This is to take care of the case where late non-root spans arrive. Instead the LowPriority decision indicates to the caller that the trace should be treated with less importance and possibly held for a lesser amount of time. In the case where a late non-root span arrives, the sub policy returning Sampled will "promote" this trace into a regular priority Pending decision"

* squash formatting change

* update internal packages to handle low priority metadata

* this commit alters logic in the main path (processor, decider) such that the processor acknowledges "LowPriority" decisions by the decider.

The cache is setup to use the tiered cache implementation. The secondary cache is hardcoded to be 10% of the size of the primary cache.

Includes tests.

fmt

* add exrta merge test to cover all lines

* use new root span policy behaviour with span_count as sub policy

* implement moviestoreguy suggestions

add t.Parallel() as per review

uses slices.Concat

fmt

use _ for zero value name

fix test

* Merged main into OBSREVE-1763-tiered-cache


Approved-by: Santosh Balaranganathan