Merged in jmoe/move-instead-of-copy (pull request #949)
refactor TraceData so it doesn't copy data and instead moves it

* refactor trace data to prevent it copying data.

The use of CopyTo() was a safety/convenience measure, and was not necessarily essential for the operation of the processor, because the TraceData object that was merged was not used after it was merged.So, a more efficient MoveTo() can be used in it's place.

This was causing 38% of the collector's overall memory allocations, so this may bring significant memory improvements, and giving the GC less to clean up.

squash this

* format

* Merged main into jmoe/move-instead-of-copy

* make the caching code block slightly more concise and clear


Approved-by: Jason Lee