Merged in jmoe/fix-secondary-promotion-eviction (pull request #968)
Jmoe/fix secondary promotion eviction

* Fixes bug which confused eviction from secondary cache with full eviction.

The trace was deleted from the secondary cache and this was treated as an eviction, not a promotion.

The issue is solved by looking at the priority of the trace metadata - if it's still low, we know it's a regular eviction. If it's regular, we know it's been promoted to the primary cache, so we don't count this as an eviction and we dont mark it as not sampled.

* fix comment

* re arrange functions as per jlee30 review


Approved-by: Santosh Balaranganathan