Merged in otelcol-113 (pull request #1052)
OBSRVE-1856 Upgrade Collector 111 -> 113

* make all-otel-latest for otelcol 0.113.0

* set default value for token sidecar endpoint, to keep tests happy

* fix deprecated usages

* remove stale comments from dockerfiles

* update running locally doc for application distro

* update config so self metrics use http not grpc

* fix override

* explicitly disable span metrics

* disable exemplars on internal metrics

* switch the internal metrics to http instead of grpc

* revert unnecessary attribute validations that were added in https://bitbucket.org/atlassian/observability-sidecar/pull-requests/922/overview

These were added due to a bug in the metrics transform processor that meant it wasn't aggregating all the labels away when `label_set: []` was configured. Now, in this collector version, they have fixed that issue and it now correctly removes all the attribtues. So, the attribute validations fail. These attributes were never supposed to be on the metric in the first place, and the information required is in the metric name.

* go mod tidy for benchmark directory. Add makefile so it's automatically tidied/linted/formatted

* make bench only run on properly tagged

* dont need state metric on cpu.utilization


Approved-by: Iris Grace Endozo