sync recent changes with upstream (#4)
regulating cache delay option which specifies the amount of time after which the processor starts regulating cache sizes. This can be used to avoid regulating cache sizes on an initial empty cache, as it can cause unnecessary cache size adjustments and high memory usage. More in docs.

Emitting span for not sampled trace. You can now specify on specific policies to emit a single span in place of the trace. This is useful if someone goes searching for a dropped trace in the system, it will tell them why it's dropped because a span appears with the policy.